requirements.txt
# FastAPI and server dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0

# Stable Diffusion and ML dependencies
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.25.0
transformers>=4.37.0
accelerate>=0.27.0
safetensors>=0.4.0

# SDXL specific enhancements
compel>=2.0.2  # For better SDXL prompt understanding

# Image processing
Pillow>=10.2.0

# HTTP client for generator script
requests>=2.31.0

# Optional: for better performance
ninja>=1.11.0

# For Windows GPU acceleration
xformers>=0.0.23 ; platform_system == "Windows"